backend/requirements.txt
flask
flask-cors
pandas
openpyxl
msal
requests
python-dotenv
